% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate_languages.R
\name{activate_languages}
\alias{activate_languages}
\title{Activates additional language}
\usage{
activate_languages(qsf, languages)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{languages}{vector of language codes}
}
\value{
qsf file as R list
}
\description{
To display a translation of a question in the survey, activate the respective language
}
\examples{
data(svy_qsf)

qsf <- activate_languages(svy_qsf, 
	languages=c("DE", "FR"))
 
 
}
