% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_skiplogic.R
\name{add_skiplogic}
\alias{add_skiplogic}
\title{Attaches a skip logic to a question}
\usage{
add_skiplogic(qsf, qid, block_id, condition, choice_id, skipto)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{qid}{question id}

\item{block_id}{block id}

\item{condition}{either "Selected", "NotSelected", "Displayed",  "NotDisplayed"}

\item{choice_id}{choice id}

\item{skipto}{either "ENDOFSURVEY" or "ENDOFBLOCK"}
}
\value{
qsf file as R list
}
\description{
Skip logic allows to send respondents to the end of the block or the end of the survey based on how they answered a question.
}
\details{
To find the ID of a question and the block ID use \code{\link{get_block_description}} and \code{\link{get_questions}}.
}
\examples{
data(svy_qsf)

# Respondents that select 'Male' (choice_id=1)
# are send to the end of the survey.

svy_mod <- add_skiplogic(svy_qsf, 
 qid='Q1',
 block_id='BL_npw6udfrvIt9oNT',
 choice_id=1,
 condition="Selected",
 skipto="ENDOFSURVEY")
 
}
