% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_question.R
\name{attach_question}
\alias{attach_question}
\title{Attaches a question to a block}
\usage{
attach_question(qsf, block_id, qid, page_break = FALSE)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{block_id}{ID of question block}

\item{qid}{QuestionID}

\item{page_break}{Add a question break? (Default: FALSE)}
}
\description{
This function appends a question to an existing survey block.
}
\examples{
data(svy_qsf)

get_block_description(svy_qsf)
qsf <- remove_block(svy_qsf, 'BL_npw6udfrvIt9oNT')
qsf <- create_block(qsf, 'demographics', id='BL_C5Z2OqNL6no9APV')

qsf <- attach_question(qsf, 
	block_id="BL_C5Z2OqNL6no9APV", 
qid=c("Q1", "Q2")) 

get_block_description(qsf)


}
