% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_displaylogic.R
\name{construct_displaylogic}
\alias{construct_displaylogic}
\title{Construct a controller for display logic}
\usage{
construct_displaylogic(...)
}
\arguments{
\item{...}{key-value tuple (see below)}
}
\description{
Helper function to construct conditions when a question is displayed.
}
\details{
Each condition comes as a tuple. The first condition takes the key "0" and all subsequent conditions are numbered. The value is a named list with the following elements: \code{qid} (QuestionID), \code{operator} ("Selected", "NotSelected"), \code{choice_id}. If there is more than one condition, all tuples other than that with key="0" must also include the parameter \code{conjuction} with the value "OR" or "AND".
}
\examples{
data(svy_qsf)

# Display question "born" only 
# when respondent selected 1 or 2 
# for the first question 

qsf_mod <- update_question(svy_qsf, 
 id_name='SecondaryAttribute', 
 id='born', 
 DisplayLogic=construct_displaylogic( 
    `0`=list(qid="Q1", operator="Selected", choice_id="1"), 
    `1`=list(qid="Q1", operator="Selected", choice_id="2", conjuction='Or'))
 )

}
\seealso{
\code{\link{update_question}}
}
