% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_validation.R
\name{construct_validation}
\alias{construct_validation}
\title{Construct a controller for force / request response feature}
\usage{
construct_validation(type = "force")
}
\arguments{
\item{type}{"force", "request", or "no" (default: "force")}
}
\description{
Helper function to enable "force a response" or "request a response" for survey questions.
}
\details{
This function is typically used in combination with \code{\link{update_question}}. See example.
}
\examples{
data(svy_qsf)

# Enable 'Request Response' validation
# for the question on respondents' sex 
# and country of birth

qsf_mod <- update_question(svy_qsf, 
 id_name='SecondaryAttribute', 
 id=c('sex', 'born'), 
 Validation=construct_validation(type='request'))


}
\seealso{
\code{\link{update_question}}
}
