% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_block.R
\name{create_block}
\alias{create_block}
\title{Create a new question block}
\usage{
create_block(qsf, description, id = NULL)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{description}{description of survey block}

\item{id}{block ID (default: random ID)}
}
\description{
Create a new question block
}
\examples{
data(svy_qsf)

get_block_description(svy_qsf)
get_block_description(
	create_block(svy_qsf, "new block"))

}
