% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_survey.R
\name{create_survey}
\alias{create_survey}
\title{Create a new Qualtrics survey file}
\usage{
create_survey(sq, page_break = FALSE)
}
\arguments{
\item{sq}{list of questions}

\item{page_break}{add page break between each question? (default: FALSE)}
}
\description{
Constructs a QSF file from a list of questions
}
\examples{
data(svy_qsf)

sq <- get_questions(svy_qsf, df=FALSE)
qsf <- create_survey(sq, page_break=TRUE)


}
