% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enable_text_entry.R
\name{enable_text_entry}
\alias{enable_text_entry}
\title{Enable "Allow text entry" for a choice}
\usage{
enable_text_entry(qsf, id_name, id, choice_id)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{id_name}{which question ID to use: "PrimaryAttribute" or "SecondaryAttribute"?}

\item{id}{Value of \code{id_name}.}

\item{choice_id}{Choice ID}
}
\description{
Enable "Allow text entry" for a choice
}
\examples{
data(svy_qsf)

enable_text_entry(svy_qsf, 
 id_name='SecondaryAttribute', 
 id='sex', choice_id=2)


}
