% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_block_description.R
\name{get_block_description}
\alias{get_block_description}
\title{Display information on question blocks}
\usage{
get_block_description(qsf)
}
\arguments{
\item{qsf}{qsf file as R list}
}
\value{
Returns a \code{tibble} with a row for each survey block, the block id, block description and the number of questions.
}
\description{
Display information on question blocks
}
\examples{
data(sq_qsf)

get_block_description(sq_qsf)

}
