% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_question.R
\name{get_question}
\alias{get_question}
\title{Extract question by ID}
\usage{
get_question(qsf, df = TRUE, id_name, id)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{df}{Return \code{data.frame}? (default: TRUE)}

\item{id_name}{which question ID to use: "PrimaryAttribute" or "SecondaryAttribute"?}

\item{id}{Value of \code{id_name}.}
}
\description{
Helper function to retrieve a single question by 
primary or secondary attribute value.
}
\examples{
data(svy_qsf)

get_question(svy_qsf, id_name='PrimaryAttribute', id='Q1')
get_question(svy_qsf, id_name='SecondaryAttribute', id='sex')


}
