% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_questions.R
\name{get_questions}
\alias{get_questions}
\title{Extract questions from QSF file}
\usage{
get_questions(qsf, df = TRUE)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{df}{convert to a \code{data.frame}? (default: TRUE)}
}
\value{
If \code{df=FALSE} the function pulls all questions from the qsf without further modifications. If \code{df=TRUE} the function returns a \code{data.frame} with columns as listed below. Note, that the conversion into a \code{data.frame} does not preserve information such as display and skip logic. 

\itemize{
	QID: Question ID (newly generated based on the sequence of questions in \code{qsf})
	ID: Choice ID (if Class="Question" then ID=0)
 Display : Displayed text 
	Class: One of three values: Question, Choices, Answers.
	Language : The language code of the question, choice or answer.
	QuestionType : Value from [['Payload']][['QuestionType']]
	PrimaryAttribute: Value from [['PrimaryAttribute']]
	SecondaryAttribute: Value from [['SecondaryAttribute']]
	TextEntry: Text-entry allowed?
	}
}
\description{
Extract questions from QSF file
}
\details{
This function extracts all questions from a QSF file and can convert those to a \code{data.frame}.
}
\examples{
data(svy_qsf)

get_questions(svy_qsf)



}
