% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qsf.R
\name{read_qsf}
\alias{read_qsf}
\title{Import a QSF file}
\usage{
read_qsf(path, ...)
}
\arguments{
\item{path}{file on disk}

\item{...}{additional conversion arguments, see \code{\link[jsonlite]{read_json}}}
}
\description{
\code{read_qsf} reads a QSF file as a list
}
\details{
This function is just a wrapper around \code{\link[jsonlite]{read_json}}. 
Any QSF file is a JSON file.
}
