% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_block.R
\name{remove_block}
\alias{remove_block}
\title{Remove a question block}
\usage{
remove_block(qsf, id)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{id}{block ID (default: random ID)}
}
\description{
Remove a question block
}
\details{
This function removes a survey block but the questions assigned to this block remain part of the survey. They can be attached to another/new survey block.
}
\examples{
data(svy_qsf)

get_block_description(svy_qsf)
get_block_description(
 remove_block(svy_qsf, id='BL_npw6udfrvIt9oNT'))


}
