% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_question.R
\name{remove_question}
\alias{remove_question}
\title{Remove a question}
\usage{
remove_question(qsf, id_name, id)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{id_name}{which question ID to use: "PrimaryAttribute" or "SecondaryAttribute"?}

\item{id}{Value of \code{id_name}.}
}
\description{
Remove a question
}
\details{
This function removes a question but leaves it attached to a question block.
}
\examples{
data(svy_qsf)

remove_question(svy_qsf, id_name='PrimaryAttribute', id='Q1')
remove_question(svy_qsf, id_name='SecondaryAttribute', id='sex')


}
