% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svy_df}
\alias{svy_df}
\title{Data frame with two survey questions}
\format{
A data frame with 12 rows and 7 variables.
\describe{
 \item{QuestionType}{Qualtrics question type}
 \item{QuestionDescription}{Question description}
 \item{QuestionID}{Question ID}
 \item{Language}{Language code}
 \item{QuestionText}{Question text}
 \item{ChoiceID}{Choice ID}
 \item{Choices}{Choice label}
}
}
\usage{
svy_df
}
\description{
A data frame representing two binary choice survey questions in three languages (English, German, and French).
}
\seealso{
\code{\link{svy_qsf}}s
}
\keyword{datasets}
