% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{svy_qsf}
\alias{svy_qsf}
\title{A qsf file as R list with two survey questions}
\format{
list object
}
\usage{
svy_qsf
}
\description{
The list object representing the qsf file includes two binary choice survey questions in three languages (English, German, and French). Includes the same questions as \code{\link{svy_df}}.
}
\seealso{
\code{\link{svy_df}}
}
\keyword{datasets}
