% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_survey.R
\name{update_survey}
\alias{update_survey}
\title{Update survey settings}
\usage{
update_survey(qsf, id_name, id, ...)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{...}{key-value tuples (see details)}
}
\description{
Update survey settings
}
\details{
The list belows catalogs some of the common keys that control the behavior of a survey. The function accepts all key-value tuples. 

\itemize{
 \item BallotBoxStuffingPrevention : 'true' or 'false' (string)
 \item SurveyTitle : Name of the survey appearing in the browser as the window or tab title. 
 \item SurveyMetaDescription : Search engines and social media services use this description. 
 \item Header : A string with JavaScript code placed in the header of the survey. 
}
}
\examples{
data(svy_qsf)

# Enable ballot box stuffing prevention 
qsf <- update_survey(qsf, 
	 BallotBoxStuffingPrevention='true')




}
