% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_qsf.R
\name{write_qsf}
\alias{write_qsf}
\title{Export survey as a QSF file}
\usage{
write_qsf(qsf, path, ...)
}
\arguments{
\item{qsf}{qsf file as R list}

\item{path}{file on disk}

\item{...}{additional conversion arguments, see \code{\link[jsonlite]{write_json}}}
}
\description{
\code{write_qsf} writes a QSF file based on a 
survey and performs some formatting checks
}
\details{
This function is just a wrapper around \code{\link[jsonlite]{write_json}}. 
Any QSF file is a JSON file.
}
\examples{
data(sq_qsf)

write_qsf(sq_qsf, path=tempfile())

}
